STEP3 함수작성 초안
